Refactorization - allow to build standalone libs
* modules are loaded by dlopen
* works currently only for Linux

+++ dxt_compress/dxt_util.c
@@ -66,7 +66,7 @@ dxt_shader_create_from_file(const char* filename, GLenum type)
     int data_size = ftell(file);
     rewind(file);
     char* program = (char*)malloc((data_size  + 1) * sizeof(char));
-    if ( data_size != fread(program, sizeof(char), data_size, file) ) {
+    if ( (size_t) data_size != fread(program, sizeof(char), data_size, file) ) {
         fprintf(stderr, "Failed to load program [%d bytes] from file %s!\n", data_size, filename);
         return 0;
     }
